fix: int64 indexing migration batch 5 - BUILD PASSING
Complete migration of int32 to int64 for indices, sizes, strides, offsets
across 16 files. Build now compiles successfully.

Core changes:
- Shape.Broadcasting.cs: All dimension/stride arrays now long[]
- NDArray.cs: Added long size constructor overloads
- TensorEngine/np.nonzero: Return NDArray<long>[] for index arrays
- ILKernelGenerator.Reduction.Axis.Simd: AxisReductionKernel delegate
  now uses long* for strides/shapes and long for sizes
- np.size: Return type changed to long
- np.array: Stride variables changed to long for pointer arithmetic
- NDArray.Indexing.Masking: Shape arrays and counts now long

Random functions:
- np.random.choice/shuffle: Added overflow checks for int.MaxValue limit
  (Random.Next only supports int; full long support deferred)

Build infrastructure:
- NumSharp.Core.csproj: Exclude *.template.cs and *.regen_disabled files

Test status: 193 failures due to memory corruption - needs investigation
in stride/offset calculations.

+++ src/NumSharp.Core/APIs/np.size.cs
@@ -11,7 +11,7 @@ namespace NumSharp
         /// <param name="axis">Axis along which the elements are counted. By default, give the total number of elements.</param>
         /// <returns>Number of elements along the specified axis.</returns>
         /// <remarks>https://numpy.org/doc/stable/reference/generated/numpy.ma.size.html</remarks>
-        public static int size(NDArray a, int? axis = null)
+        public static long size(NDArray a, int? axis = null)
         {
             if (a == null)
                 throw new ArgumentNullException(nameof(a));
