+++ src/NumSharp.Core/LinearAlgebra/NDArray.matrix_power.cs
@@ -14,7 +14,8 @@ namespace NumSharp
             for (int idx = 2; idx <= power; idx++)
                 product = TensorEngine.Dot(product, this);
 
-            product = (power == 0) ? np.eye(product.shape[0]) : product;
+            // Matrix dimensions are typically small; np.eye expects int
+            product = (power == 0) ? np.eye((int)product.shape[0]) : product;
 
             return product;
         }
